* origin/main: (32 commits)
  Engine version checks no longer needed for new engines (#4843)
  fix: wait for new balance on transfer asset (#4836)
  chore: re-enable try-runtime tests in upgrade-test (#4840)
  Feat: per chain safe mode to restrict deposits (#4819)
  fix: upgrade test permissions (#4838)
  Update Arbitrum migrations and chainspec (#4835)
  chore: print correct logs in upgrade-test (#4832)
  feat: Affiliate Brokers (#4777)
  chore: resolve build warnings (#4831)
  chore: remove old dylibs (#4830)
  feat: Efficient Arbitrum witnessing (PRO-1098) (#4811)
  chore: change CI command (#4829)
  fix(docker images): dont delete `/sbin` dirs 😵‍💫 (#4828)
  fix: improve container pull times ⌚️ (#4827)
  refactor(bouncer): refactor setup (#4825)
  Realistic BTC egress fees (#4822)
  chore: inline and functional split and group (#4808)
  feat: LP Asset rebalancing (PRO-1259) (#4784)
  feat: store boost fees explicitly + rpc (#4818)
  feat: Governance extrinsic to create boost pools (#4816)
  ...

# Conflicts:
#	Cargo.lock
#	state-chain/chains/src/sol.rs
#	state-chain/runtime/src/migrations/arbitrum_integration.rs
#	state-chain/runtime/src/safe_mode.rs

+++ engine/src/witness/eth.rs
@@ -123,7 +123,7 @@ where
 	tracing::info!("Safety margin for Ethereum is set to {eth_safety_margin} blocks.",);
 
 	let eth_safe_vault_source = eth_source
-		.lag_safety(eth_safety_margin as usize)
+		.lag_safety(eth_safety_margin)
 		.logging("safe block produced")
 		.chunk_by_vault(vaults, scope);
